requirements.txt
asgiref==3.4.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
certifi==2021.5.30
charset-normalizer==2.0.6
coverage @ file:///tmp/build/80754af9/coverage_1614614861224/work
cycler==0.10.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1627074853231/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
Django==3.2.8
djangorestframework==3.12.4
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
idna==3.2
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1631291098355/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1632763773116/work
ipython-genutils==0.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1610146787869/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1631852705892/work
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
more-itertools @ file:///tmp/build/80754af9/more-itertools_1622818384463/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264704411/work
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandas==1.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.3.2
pluggy==1.0.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1629903925368/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py @ file:///tmp/build/80754af9/py_1607971587848/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pytest==6.2.5
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1622038437086/work
requests==2.26.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094708661/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1630423529112/work
urllib3==1.26.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
